requirements.txt
numpy==2.3.3
scipy==1.16.2
matplotlib==3.10.6
py-pde==0.46.0
numba==0.62.0
llvmlite==0.45.0
h5py==3.14.0
tqdm==4.67.1
SciencePlots==2.1.1
contourpy==1.3.3
cycler==0.12.1
kiwisolver==1.4.9
pillow==11.3.0
sympy==1.14.0
mpmath==1.3.0
python-dateutil==2.9.0.post0
pyparsing==3.2.5
fonttools==4.60.0
packaging==25.0
six==1.17.0